refine the bert train  with updated common func
