+++ conf/denas/cv/e2eaiok_denas_train_cnn.conf
@@ -5,7 +5,7 @@ eval_epochs: 1
 input_size: 32
 best_model_structure: /home/vmagent/app/e2eAIOK/e2eAIOK/DeNas/best_model_structure.txt
 num_classes: 10
-dist_backend: ccl
+dist_backend: gloo
 train_batch_size: 128
 eval_batch_size: 128
 data_path: ~/data/pytorch_cifar10
