+++ conf/denas/nlp/e2eaiok_denas_train_bert.conf
@@ -1,13 +1,37 @@
 # conf for Bert-based DE-NAS
+domain: bert
 task_name: squad1
-arches_file: /home/vmagent/app/e2eaiok/DeNas/
-model: home/vmagent/app/dataset/bert-base-uncased/
+data_set: SQuADv1.1
+num_train_examples: 87599
+best_model_structure: /home/vmagent/app/e2eAIOK/e2eAIOK/DeNas/best_model_structure.txt
+model: /home/vmagent/app/dataset/bert-base-uncased/
 model_dir: /home/vmagent/app/dataset/bert-base-uncased/
 data_dir: /home/vmagent/app/dataset/SQuAD/
-output_dir: /home/vmagent/app/e2eaiok/DeNas/nlp/output/
-num_train_epochs: '4'
-max_seq_length: '384'
-learning_rate: '3e-5'
-train_batch_size: '12'
-eval_step: '200'
-f1_threshold: '81.5'
+output_dir: /home/vmagent/app/e2eAIOK/e2eAIOK/DeNas/nlp/
+dist_backend: gloo
+gradient_accumulation_steps: 1
+warmup_proportion: 0.1
+learning_rate: 0.00003
+weight_decay: 0.0001
+train_epochs: 4
+max_seq_length: 384
+doc_stride: 128
+train_batch_size: 12
+eval_batch_size: 32
+eval_step: 200
+n_best_size: 20
+max_answer_length: 30
+max_query_length: 64
+criterion: "CrossEntropyQALoss"
+optimizer: "BertAdam"
+lr_scheduler: "warmup_linear"
+version_2_with_negative: 0
+null_score_diff_threshold: 0.0
+num_labels: 2
+num_workers: 10
+pin_mem: True
+verbose_logging: True
+no_cuda: True
+do_lower_case: True
+metric_threshold: 81.5
+eval_metric: "qa_f1"
