+++ e2eAIOK/DeNas/module/nlp/bert_encoder_super.py
@@ -6,10 +6,10 @@ import copy
 import torch
 from torch import nn
 
-from module.nlp.Linear_super import LinearSuper as SuperLinear
-from module.nlp.layernorm_super import LayerNormSuper as SuperBertLayerNorm
-from module.nlp.bert_attention_super import SuperBertAttention
-from module.nlp.bert_intermediate_super import SuperBertIntermediate
+from e2eAIOK.DeNas.module.nlp.Linear_super import LinearSuper as SuperLinear
+from e2eAIOK.DeNas.module.nlp.layernorm_super import LayerNormSuper as SuperBertLayerNorm
+from e2eAIOK.DeNas.module.nlp.bert_attention_super import SuperBertAttention
+from e2eAIOK.DeNas.module.nlp.bert_intermediate_super import SuperBertIntermediate
 
 
 logging.basicConfig(format='%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
