+++ e2eAIOK/DeNas/module/nlp/layernorm_super.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from module.layernorm_base import LayerNormBase
+from e2eAIOK.DeNas.module.layernorm_base import LayerNormBase
 
 
 class LayerNormSuper(LayerNormBase):
