--- e2eAIOK/DeNas/nlp/supernet_train.py
@@ -1,1605 +1,0 @@
-# coding=utf-8
-# Copyright (c) 2022, Intel. and its affiliates.
-# 2021.12.30-Changed for TinyBERT general distillation
-#      Huawei Technologies Co., Ltd. <yinyichun@huawei.com>
-# Copyright 2021 Huawei Technologies Co., Ltd.
-# Copyright 2018 The Google AI Language Team Authors, The HuggingFace Inc. team.
-# Copyright (c) 2018, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""BERT finetuning runner."""
-
-from __future__ import absolute_import, division, print_function
-
-import ast
-import os
-import re
-import sys
-import csv
-import math
-import json
-import string
-import argparse
-import logging
-import random
-import collections
-
-import time
-from datetime import datetime
-from tqdm import tqdm, trange
-from scipy.stats import pearsonr, spearmanr
-from sklearn.metrics import matthews_corrcoef, f1_score
-from collections import Counter
-import numpy as np
-
-import torch
-from torch.utils.data import (DataLoader, RandomSampler, SequentialSampler, TensorDataset)
-from torch.nn import CrossEntropyLoss, MSELoss
-
-
-from supernet_bert import SuperBertForQuestionAnswering, BertConfig
-from module.nlp.tokenization import BertTokenizer, BasicTokenizer, whitespace_tokenize
-from module.nlp.optimization import BertAdam
-from nlp.utils import get_subconfig
-
-
-csv.field_size_limit(sys.maxsize)
-
-logging.basicConfig(format='%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
-                    datefmt='%m/%d/%Y %H:%M:%S',
-                    level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-OUTPUT_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)), "output")
-MODEL_DIR = "/home/vmagent/app/dataset/bert-base-uncased/"
-DATA_DIR = "/home/vmagent/app/dataset/SQuAD"
-
-class InputExample(object):
-    """A single training/test example for simple sequence classification."""
-
-    def __init__(self, guid, text_a, text_b=None, label=None):
-        """Constructs a InputExample.
-
-        Args:
-            guid: Unique id for the example.
-            text_a: string. The untokenized text of the first sequence. For single
-            sequence tasks, only this sequence must be specified.
-            text_b: (Optional) string. The untokenized text of the second sequence.
-            Only must be specified for sequence pair tasks.
-            label: (Optional) string. The label of the example. This should be
-            specified for train and dev examples, but not for test examples.
-        """
-        self.guid = guid
-        self.text_a = text_a
-        self.text_b = text_b
-        self.label = label
-
-
-class SquadExample(object):
-    def __init__(self,
-                 qas_id,
-                 question_text,
-                 doc_tokens,
-                 orig_answer_text=None,
-                 start_position=None,
-                 end_position=None,
-                 is_impossible=None):
-        self.qas_id = qas_id
-        self.question_text = question_text
-        self.doc_tokens = doc_tokens
-        self.orig_answer_text = orig_answer_text
-        self.start_position = start_position
-        self.end_position = end_position
-        self.is_impossible = is_impossible
-
-    def __str__(self):
-        return self.__repr__()
-
-    def __repr__(self):
-        s = ""
-        s += "qas_id: %s" % (self.qas_id)
-        s += ", question_text: %s" % (
-            self.question_text)
-        s += ", doc_tokens: [%s]" % (" ".join(self.doc_tokens))
-        if self.start_position:
-            s += ", start_position: %d" % (self.start_position)
-        if self.end_position:
-            s += ", end_position: %d" % (self.end_position)
-        if self.is_impossible:
-            s += ", is_impossible: %r" % (self.is_impossible)
-        return s
-
-
-class InputFeatures(object):
-    """A single set of features of data."""
-
-    def __init__(self, input_ids, input_mask, segment_ids, label_id, seq_length=None):
-        self.input_ids = input_ids
-        self.input_mask = input_mask
-        self.segment_ids = segment_ids
-        self.seq_length = seq_length
-        self.label_id = label_id
-
-
-class QAInputFeatures(object):
-    """A single set of features of data."""
-
-    def __init__(self,
-                 unique_id,
-                 example_index,
-                 doc_span_index,
-                 tokens,
-                 token_to_orig_map,
-                 token_is_max_context,
-                 input_ids,
-                 input_mask,
-                 segment_ids,
-                 start_position=None,
-                 end_position=None,
-                 is_impossible=None):
-        self.unique_id = unique_id
-        self.example_index = example_index
-        self.doc_span_index = doc_span_index
-        self.tokens = tokens
-        self.token_to_orig_map = token_to_orig_map
-        self.token_is_max_context = token_is_max_context
-        self.input_ids = input_ids
-        self.input_mask = input_mask
-        self.segment_ids = segment_ids
-        self.start_position = start_position
-        self.end_position = end_position
-        self.is_impossible = is_impossible
-
-
-class DataProcessor(object):
-    """Base class for data converters for sequence classification data sets."""
-
-    def get_train_examples(self, data_dir):
-        """Gets a collection of `InputExample`s for the train set."""
-        raise NotImplementedError()
-
-    def get_dev_examples(self, data_dir):
-        """Gets a collection of `InputExample`s for the dev set."""
-        raise NotImplementedError()
-
-    def get_labels(self):
-        """Gets the list of labels for this data set."""
-        raise NotImplementedError()
-
-    @classmethod
-    def _read_tsv(cls, input_file, quotechar=None):
-        """Reads a tab separated value file."""
-        with open(input_file, "r", encoding="utf-8") as f:
-            reader = csv.reader(f, delimiter="\t", quotechar=quotechar)
-            lines = []
-            for line in reader:
-                if sys.version_info[0] == 2:
-                    line = list(unicode(cell, 'utf-8') for cell in line)
-                lines.append(line)
-            return lines
-
-
-class Squad1Processor(DataProcessor):
-
-    def get_train_examples(self, data_dir):
-        """See base class."""
-        return self._create_examples(
-            self._read_tsv(os.path.join(data_dir, "train.tsv")), "train")
-
-    def get_dev_examples(self, data_dir):
-        """See base class."""
-        return self._create_examples(
-            self._read_tsv(os.path.join(data_dir, "dev.tsv")), "dev")
-
-    def get_labels(self):
-        """See base class."""
-        return ["0", "1"]
-
-    def _create_examples(self, lines, set_type):
-        """Creates examples for the training and dev sets."""
-        examples = []
-        for (i, line) in enumerate(lines):
-            if i == 0:
-                continue
-            guid = "%s-%s" % (set_type, line[0])
-            text_a = line[1]
-            text_b = line[2]
-            label = line[-1]
-            examples.append(
-                InputExample(guid=guid, text_a=text_a, text_b=text_b, label=label))
-        return examples
-
-
-def _improve_answer_span(doc_tokens, input_start, input_end, tokenizer,
-                         orig_answer_text):
-    """Returns tokenized answer spans that better match the annotated answer."""
-    tok_answer_text = " ".join(tokenizer.tokenize(orig_answer_text))
-
-    for new_start in range(input_start, input_end + 1):
-        for new_end in range(input_end, new_start - 1, -1):
-            text_span = " ".join(doc_tokens[new_start:(new_end + 1)])
-            if text_span == tok_answer_text:
-                return (new_start, new_end)
-
-    return (input_start, input_end)
-
-
-def _check_is_max_context(doc_spans, cur_span_index, position):
-    """Check if this is the 'max context' doc span for the token."""
-    best_score = None
-    best_span_index = None
-    for (span_index, doc_span) in enumerate(doc_spans):
-        end = doc_span.start + doc_span.length - 1
-        if position < doc_span.start:
-            continue
-        if position > end:
-            continue
-        num_left_context = position - doc_span.start
-        num_right_context = end - position
-        score = min(num_left_context, num_right_context) + 0.01 * doc_span.length
-        if best_score is None or score > best_score:
-            best_score = score
-            best_span_index = span_index
-
-    return cur_span_index == best_span_index
-
-
-def qa_convert_examples_to_features(examples, tokenizer, max_seq_length,
-                                    doc_stride, max_query_length,
-                                    is_training):
-    """Loads a data file into a list of `InputBatch`s."""
-    question_part_length_list = []
-    passage_part_length_list = []
-
-    unique_id = 1000000000
-
-    features = []
-    for (example_index, example) in enumerate(examples):
-        query_tokens = tokenizer.tokenize(example.question_text)
-
-        if len(query_tokens) > max_query_length:
-            query_tokens = query_tokens[0:max_query_length]
-
-        tok_to_orig_index = []
-        orig_to_tok_index = []
-        all_doc_tokens = []
-        for (i, token) in enumerate(example.doc_tokens):
-            orig_to_tok_index.append(len(all_doc_tokens))
-            sub_tokens = tokenizer.tokenize(token)
-            for sub_token in sub_tokens:
-                tok_to_orig_index.append(i)
-                all_doc_tokens.append(sub_token)
-
-        tok_start_position = None
-        tok_end_position = None
-        if is_training and example.is_impossible:
-            tok_start_position = -1
-            tok_end_position = -1
-        if is_training and not example.is_impossible:
-            tok_start_position = orig_to_tok_index[example.start_position]
-            if example.end_position < len(example.doc_tokens) - 1:
-                tok_end_position = orig_to_tok_index[example.end_position + 1] - 1
-            else:
-                tok_end_position = len(all_doc_tokens) - 1
-            (tok_start_position, tok_end_position) = _improve_answer_span(
-                all_doc_tokens, tok_start_position, tok_end_position, tokenizer,
-                example.orig_answer_text)
-
-        # The -3 accounts for [CLS], [SEP] and [SEP]
-        max_tokens_for_doc = max_seq_length - len(query_tokens) - 3
-
-        # We can have documents that are longer than the maximum sequence length.
-        # To deal with this we do a sliding window approach, where we take chunks
-        # of the up to our max length with a stride of `doc_stride`.
-        _DocSpan = collections.namedtuple(  # pylint: disable=invalid-name
-            "DocSpan", ["start", "length"])
-        doc_spans = []
-        start_offset = 0
-        while start_offset < len(all_doc_tokens):
-            length = len(all_doc_tokens) - start_offset
-            if length > max_tokens_for_doc:
-                length = max_tokens_for_doc
-            doc_spans.append(_DocSpan(start=start_offset, length=length))
-            if start_offset + length == len(all_doc_tokens):
-                break
-            start_offset += min(length, doc_stride)
-
-        for (doc_span_index, doc_span) in enumerate(doc_spans):
-            tokens = []
-            token_to_orig_map = {}
-            token_is_max_context = {}
-            segment_ids = []
-            tokens.append("[CLS]")
-            segment_ids.append(0)
-            for token in query_tokens:
-                tokens.append(token)
-                segment_ids.append(0)
-            tokens.append("[SEP]")
-            segment_ids.append(0)
-            question_part_length_list.append(len(tokens))                                    # 记录question的长度
-
-            for i in range(doc_span.length):
-                split_token_index = doc_span.start + i
-                token_to_orig_map[len(tokens)] = tok_to_orig_index[split_token_index]
-
-                is_max_context = _check_is_max_context(doc_spans, doc_span_index,
-                                                       split_token_index)
-                token_is_max_context[len(tokens)] = is_max_context
-                tokens.append(all_doc_tokens[split_token_index])
-                segment_ids.append(1)
-            tokens.append("[SEP]")
-            segment_ids.append(1)
-
-            passage_part_length_list.append(len(tokens))                                    # 记录question + passage的长度
-
-            input_ids = tokenizer.convert_tokens_to_ids(tokens)
-
-            # The mask has 1 for real tokens and 0 for padding tokens. Only real
-            # tokens are attended to.
-            input_mask = [1] * len(input_ids)
-
-            # Zero-pad up to the sequence length.
-            while len(input_ids) < max_seq_length:
-                input_ids.append(0)
-                input_mask.append(0)
-                segment_ids.append(0)
-
-            assert len(input_ids) == max_seq_length
-            assert len(input_mask) == max_seq_length
-            assert len(segment_ids) == max_seq_length
-
-            start_position = None
-            end_position = None
-            if is_training and not example.is_impossible:
-                # For training, if our document chunk does not contain an annotation
-                # we throw it out, since there is nothing to predict.
-                doc_start = doc_span.start
-                doc_end = doc_span.start + doc_span.length - 1
-                out_of_span = False
-                if not (tok_start_position >= doc_start and
-                        tok_end_position <= doc_end):
-                    out_of_span = True
-                if out_of_span:
-                    start_position = 0
-                    end_position = 0
-                else:
-                    doc_offset = len(query_tokens) + 2
-                    start_position = tok_start_position - doc_start + doc_offset         # 确实是变成整个input的位置
-                    end_position = tok_end_position - doc_start + doc_offset
-            if is_training and example.is_impossible:
-                start_position = 0
-                end_position = 0
-
-            features.append(
-                QAInputFeatures(
-                    unique_id=unique_id,
-                    example_index=example_index,
-                    doc_span_index=doc_span_index,
-                    tokens=tokens,
-                    token_to_orig_map=token_to_orig_map,
-                    token_is_max_context=token_is_max_context,
-                    input_ids=input_ids,
-                    input_mask=input_mask,
-                    segment_ids=segment_ids,
-                    start_position=start_position,
-                    end_position=end_position,
-                    is_impossible=example.is_impossible))
-            unique_id += 1
-
-    return features
-
-
-def convert_examples_to_features(examples, label_list, max_seq_length,
-                                 tokenizer, output_mode):
-    """Loads a data file into a list of `InputBatch`s."""
-
-    label_map = {label: i for i, label in enumerate(label_list)}
-
-    features = []
-    for (ex_index, example) in enumerate(examples):
-        if ex_index % 10000 == 0:
-            logger.info("Writing example %d of %d" % (ex_index, len(examples)))
-
-        tokens_a = tokenizer.tokenize(example.text_a)
-
-        tokens_b = None
-        if example.text_b:
-            tokens_b = tokenizer.tokenize(example.text_b)
-            _truncate_seq_pair(tokens_a, tokens_b, max_seq_length - 3)
-        else:
-            if len(tokens_a) > max_seq_length - 2:
-                tokens_a = tokens_a[:(max_seq_length - 2)]
-
-        tokens = ["[CLS]"] + tokens_a + ["[SEP]"]
-        segment_ids = [0] * len(tokens)
-
-        if tokens_b:
-            tokens += tokens_b + ["[SEP]"]
-            segment_ids += [1] * (len(tokens_b) + 1)
-
-        input_ids = tokenizer.convert_tokens_to_ids(tokens)
-        input_mask = [1] * len(input_ids)
-        seq_length = len(input_ids)
-
-        padding = [0] * (max_seq_length - len(input_ids))
-        input_ids += padding
-        input_mask += padding
-        segment_ids += padding
-
-        assert len(input_ids) == max_seq_length
-        assert len(input_mask) == max_seq_length
-        assert len(segment_ids) == max_seq_length
-
-        if output_mode == "classification":
-            label_id = label_map[example.label]
-        elif output_mode == "regression":
-            label_id = float(example.label)
-        else:
-            raise KeyError(output_mode)
-
-        if ex_index < 1:
-            logger.info("*** Example ***")
-            logger.info("guid: %s" % (example.guid))
-            logger.info("tokens: %s" % " ".join(
-                [str(x) for x in tokens]))
-            logger.info("input_ids: %s" % " ".join([str(x) for x in input_ids]))
-            logger.info("input_mask: %s" % " ".join([str(x) for x in input_mask]))
-            logger.info(
-                "segment_ids: %s" % " ".join([str(x) for x in segment_ids]))
-            logger.info("label: {}".format(example.label))
-            logger.info("label_id: {}".format(label_id))
-
-        features.append(
-            InputFeatures(input_ids=input_ids,
-                          input_mask=input_mask,
-                          segment_ids=segment_ids,
-                          label_id=label_id,
-                          seq_length=seq_length))
-    return features
-
-
-def read_squad_examples(input_file, is_training, version_2_with_negative):
-    with open(input_file, "r", encoding='utf-8') as reader:
-        input_data = json.load(reader)["data"]
-
-    def is_whitespace(c):
-        if c == " " or c == "\t" or c == "\r" or c == "\n" or ord(c) == 0x202F:
-            return True
-        return False
-
-    examples = []
-    for entry in input_data:
-        for paragraph in entry["paragraphs"]:
-            paragraph_text = paragraph["context"]
-            doc_tokens = []
-            char_to_word_offset = []
-            prev_is_whitespace = True
-            for c in paragraph_text:
-                if is_whitespace(c):
-                    prev_is_whitespace = True
-                else:
-                    if prev_is_whitespace:
-                        doc_tokens.append(c)
-                    else:
-                        doc_tokens[-1] += c
-                    prev_is_whitespace = False
-                char_to_word_offset.append(len(doc_tokens) - 1)
-
-            for qa in paragraph["qas"]:
-
-                qas_id = qa["id"]
-                question_text = qa["question"]
-                start_position = None
-                end_position = None
-                orig_answer_text = None
-                is_impossible = False
-                if is_training:
-                    if version_2_with_negative:
-                        if 'is_impossible' not in qa:
-                            qa['is_impossible'] = True
-                        is_impossible = qa["is_impossible"]
-                    if (len(qa["answers"]) != 1) and (not is_impossible):
-                        raise ValueError(
-                            "For training, each question should have exactly 1 answer.")
-                    if not is_impossible:
-                        answer = qa["answers"][0]
-                        orig_answer_text = answer["text"]
-                        answer_offset = answer["answer_start"]
-                        answer_length = len(orig_answer_text)
-                        start_position = char_to_word_offset[answer_offset]
-                        end_position = char_to_word_offset[answer_offset + answer_length - 1]
-                        # Only add answers where the text can be exactly recovered from the
-                        # document. If this CAN'T happen it's likely due to weird Unicode
-                        # stuff so we will just skip the example.
-                        #
-                        # Note that this means for training mode, every example is NOT
-                        # guaranteed to be preserved.
-                        actual_text = " ".join(doc_tokens[start_position:(end_position + 1)])
-                        cleaned_answer_text = " ".join(
-                            whitespace_tokenize(orig_answer_text))
-                        if actual_text.find(cleaned_answer_text) == -1:
-                            logger.warning("Could not find answer")
-                            continue
-                    else:
-                        start_position = -1
-                        end_position = -1
-                        orig_answer_text = ""
-
-                example = SquadExample(
-                    qas_id=qas_id,
-                    question_text=question_text,
-                    doc_tokens=doc_tokens,
-                    orig_answer_text=orig_answer_text,
-                    start_position=start_position,
-                    end_position=end_position,
-                    is_impossible=is_impossible)
-                examples.append(example)
-    logger.info('load {} examples!'.format(len(examples)))
-    return input_data, examples
-
-
-def _truncate_seq_pair(tokens_a, tokens_b, max_length):
-    """Truncates a sequence pair in place to the maximum length."""
-    while True:
-        total_length = len(tokens_a) + len(tokens_b)
-        if total_length <= max_length:
-            break
-        if len(tokens_a) > len(tokens_b):
-            tokens_a.pop()
-        else:
-            tokens_b.pop()
-
-
-def _f1_score(prediction, ground_truth):
-    prediction_tokens = normalize_answer(prediction).split()
-    ground_truth_tokens = normalize_answer(ground_truth).split()
-    common = Counter(prediction_tokens) & Counter(ground_truth_tokens)
-    num_same = sum(common.values())
-    if num_same == 0:
-        return 0
-    precision = 1.0 * num_same / len(prediction_tokens)
-    recall = 1.0 * num_same / len(ground_truth_tokens)
-    f1 = (2 * precision * recall) / (precision + recall)
-    return f1
-
-
-def simple_accuracy(preds, labels):
-    return (preds == labels).mean()
-
-
-def acc_and_f1(preds, labels):
-    acc = simple_accuracy(preds, labels)
-    f1 = f1_score(y_true=labels, y_pred=preds)
-    return {
-        "acc": acc,
-        "f1": f1,
-        "acc_and_f1": (acc + f1) / 2,
-    }
-
-
-def pearson_and_spearman(preds, labels):
-    pearson_corr = pearsonr(preds, labels)[0]
-    spearman_corr = spearmanr(preds, labels)[0]
-    return {
-        "pearson": pearson_corr,
-        "spearmanr": spearman_corr,
-        "corr": (pearson_corr + spearman_corr) / 2,
-    }
-
-
-def compute_metrics(task_name, preds, labels):
-    assert len(preds) == len(labels)
-    if task_name == "cola":
-        return {"mcc": matthews_corrcoef(labels, preds)}
-    elif task_name == "sst-2":
-        return {"acc": simple_accuracy(preds, labels)}
-    elif task_name == "mrpc":
-        return acc_and_f1(preds, labels)
-    elif task_name == "sts-b":
-        return pearson_and_spearman(preds, labels)
-    elif task_name == "qqp":
-        return acc_and_f1(preds, labels)
-    elif task_name == "mnli":
-        return {"acc": simple_accuracy(preds, labels)}
-    elif task_name == "mnli-mm":
-        return {"acc": simple_accuracy(preds, labels)}
-    elif task_name == "qnli":
-        return {"acc": simple_accuracy(preds, labels)}
-    elif task_name == "rte":
-        return {"acc": simple_accuracy(preds, labels)}
-    elif task_name == "wnli":
-        return {"acc": simple_accuracy(preds, labels)}
-    else:
-        raise KeyError(task_name)
-
-
-def get_tensor_data(output_mode, features):
-    if output_mode == "classification":
-        all_label_ids = torch.tensor([f.label_id for f in features], dtype=torch.long)
-    elif output_mode == "regression":
-        all_label_ids = torch.tensor([f.label_id for f in features], dtype=torch.float)
-
-    all_seq_lengths = torch.tensor([f.seq_length for f in features], dtype=torch.long)
-    all_input_ids = torch.tensor([f.input_ids for f in features], dtype=torch.long)
-    all_input_mask = torch.tensor([f.input_mask for f in features], dtype=torch.long)
-    all_segment_ids = torch.tensor([f.segment_ids for f in features], dtype=torch.long)
-    tensor_data = TensorDataset(all_input_ids, all_input_mask, all_segment_ids,
-                                all_label_ids, all_seq_lengths)
-    return tensor_data, all_label_ids
-
-
-def result_to_file(result, file_name):
-    with open(file_name, "a") as writer:
-        logger.info("***** Eval results *****")
-        for key in sorted(result.keys()):
-            logger.info("  %s = %s", key, str(result[key]))
-            writer.write("%s = %s\n" % (key, str(result[key])))
-
-
-def do_eval(model, task_name, eval_dataloader,
-            device, output_mode, eval_labels, num_labels, subbert_config):
-    eval_loss = 0
-    nb_eval_steps = 0
-    preds = []
-    infer_times = []
-    for batch_ in tqdm(eval_dataloader, desc="Evaluating"):
-        batch_ = tuple(t.to(device) for t in batch_)
-        with torch.no_grad():
-            input_ids, input_mask, segment_ids, label_ids, seq_lengths = batch_
-            start = datetime.now()
-            logits = model(input_ids, subbert_config, input_mask, segment_ids)
-            infer_times.append((datetime.now() - start).microseconds / 1000)
-        # create eval loss and other metric required by the task
-        if output_mode == "classification":
-            loss_fct = CrossEntropyLoss()
-            tmp_eval_loss = loss_fct(logits.view(-1, num_labels), label_ids.view(-1))
-        elif output_mode == "regression":
-            loss_fct = MSELoss()
-            tmp_eval_loss = loss_fct(logits.view(-1), label_ids.view(-1))
-
-        eval_loss += tmp_eval_loss.mean().item()
-        nb_eval_steps += 1
-        if len(preds) == 0:
-            preds.append(logits.detach().cpu().numpy())
-        else:
-            preds[0] = np.append(
-                preds[0], logits.detach().cpu().numpy(), axis=0)
-
-    eval_loss = eval_loss / nb_eval_steps
-
-    preds = preds[0]
-    if output_mode == "classification":
-        preds = np.argmax(preds, axis=1)
-    elif output_mode == "regression":
-        preds = np.squeeze(preds)
-    result = compute_metrics(task_name, preds, eval_labels.numpy())
-    result['eval_loss'] = eval_loss
-    result['infer_cnt'] = len(infer_times)
-    result['infer_time'] = sum(infer_times) / len(infer_times)
-    logger.info("******** Eval results ********")
-    for key in sorted(result.keys()):
-        logger.info(" dev: %s = %s", key, str(result[key]))
-    return result
-
-
-RawResult = collections.namedtuple("RawResult",
-                                   ["unique_id", "start_logits", "end_logits"])
-
-
-def _get_best_indexes(logits, n_best_size):
-    """Get the n-best logits from a list."""
-    index_and_score = sorted(enumerate(logits), key=lambda x: x[1], reverse=True)
-
-    best_indexes = []
-    for i in range(len(index_and_score)):
-        if i >= n_best_size:
-            break
-        best_indexes.append(index_and_score[i][0])
-    return best_indexes
-
-
-def get_final_text(pred_text, orig_text, do_lower_case, verbose_logging=False):
-    """Project the tokenized prediction back to the original text."""
-    def _strip_spaces(text):
-        ns_chars = []
-        ns_to_s_map = collections.OrderedDict()
-        for (i, c) in enumerate(text):
-            if c == " ":
-                continue
-            ns_to_s_map[len(ns_chars)] = i
-            ns_chars.append(c)
-        ns_text = "".join(ns_chars)
-        return (ns_text, ns_to_s_map)
-
-    # We first tokenize `orig_text`, strip whitespace from the result
-    # and `pred_text`, and check if they are the same length. If they are
-    # NOT the same length, the heuristic has failed. If they are the same
-    # length, we assume the characters are one-to-one aligned.
-    tokenizer = BasicTokenizer(do_lower_case=do_lower_case)
-
-    tok_text = " ".join(tokenizer.tokenize(orig_text))
-
-    start_position = tok_text.find(pred_text)
-    if start_position == -1:
-        if verbose_logging:
-            logger.info(
-                "Unable to find text: '%s' in '%s'" % (pred_text, orig_text))
-        return orig_text
-    end_position = start_position + len(pred_text) - 1
-
-    (orig_ns_text, orig_ns_to_s_map) = _strip_spaces(orig_text)
-    (tok_ns_text, tok_ns_to_s_map) = _strip_spaces(tok_text)
-
-    if len(orig_ns_text) != len(tok_ns_text):
-        if verbose_logging:
-            logger.info("Length not equal after stripping spaces: '%s' vs '%s'",
-                        orig_ns_text, tok_ns_text)
-        return orig_text
-
-    # We then project the characters in `pred_text` back to `orig_text` using
-    # the character-to-character alignment.
-    tok_s_to_ns_map = {}
-    for (i, tok_index) in tok_ns_to_s_map.items():
-        tok_s_to_ns_map[tok_index] = i
-
-    orig_start_position = None
-    if start_position in tok_s_to_ns_map:
-        ns_start_position = tok_s_to_ns_map[start_position]
-        if ns_start_position in orig_ns_to_s_map:
-            orig_start_position = orig_ns_to_s_map[ns_start_position]
-
-    if orig_start_position is None:
-        if verbose_logging:
-            logger.info("Couldn't map start position")
-        return orig_text
-
-    orig_end_position = None
-    if end_position in tok_s_to_ns_map:
-        ns_end_position = tok_s_to_ns_map[end_position]
-        if ns_end_position in orig_ns_to_s_map:
-            orig_end_position = orig_ns_to_s_map[ns_end_position]
-
-    if orig_end_position is None:
-        if verbose_logging:
-            logger.info("Couldn't map end position")
-        return orig_text
-
-    output_text = orig_text[orig_start_position:(orig_end_position + 1)]
-    return output_text
-
-
-def _compute_softmax(scores):
-    """Compute softmax probability over raw logits."""
-    if not scores:
-        return []
-
-    max_score = None
-    for score in scores:
-        if max_score is None or score > max_score:
-            max_score = score
-
-    exp_scores = []
-    total_sum = 0.0
-    for score in scores:
-        x = math.exp(score - max_score)
-        exp_scores.append(x)
-        total_sum += x
-
-    probs = []
-    for score in exp_scores:
-        probs.append(score / total_sum)
-    return probs
-
-
-def normalize_answer(s):
-    """Lower text and remove punctuation, articles and extra whitespace."""
-    def remove_articles(text):
-        regex = re.compile(r'\b(a|an|the)\b', re.UNICODE)
-        return re.sub(regex, ' ', text)
-    def white_space_fix(text):
-        return ' '.join(text.split())
-    def remove_punc(text):
-        exclude = set(string.punctuation)
-        return ''.join(ch for ch in text if ch not in exclude)
-    def lower(text):
-        return text.lower()
-    return white_space_fix(remove_articles(remove_punc(lower(s))))
-
-
-def get_tokens(s):
-    if not s: return []
-    return normalize_answer(s).split()
-
-
-def make_eval_dict(exact_scores, f1_scores, qid_list=None):
-    if not qid_list:
-        total = len(exact_scores)
-        return collections.OrderedDict([
-            ('exact_match', 100.0 * sum(exact_scores.values()) / total),
-            ('f1', 100.0 * sum(f1_scores.values()) / total),
-            ('total', total),
-        ])
-    else:
-        total = len(qid_list)
-        return collections.OrderedDict([
-            ('exact_match', 100.0 * sum(exact_scores[k] for k in qid_list) / total),
-            ('f1', 100.0 * sum(f1_scores[k] for k in qid_list) / total),
-            ('total', total),
-        ])
-
-
-def compute_exact(a_gold, a_pred):
-    return int(normalize_answer(a_gold) == normalize_answer(a_pred))
-
-
-def compute_f1(a_gold, a_pred):
-    gold_toks = get_tokens(a_gold)
-    pred_toks = get_tokens(a_pred)
-    common = collections.Counter(gold_toks) & collections.Counter(pred_toks)
-    num_same = sum(common.values())
-    if len(gold_toks) == 0 or len(pred_toks) == 0:
-        # If either is no-answer, then F1 is 1 if they agree, 0 otherwise
-        return int(gold_toks == pred_toks)
-    if num_same == 0:
-        return 0
-    precision = 1.0 * num_same / len(pred_toks)
-    recall = 1.0 * num_same / len(gold_toks)
-    f1 = (2 * precision * recall) / (precision + recall)
-    return f1
-
-
-def apply_no_ans_threshold(scores, na_probs, qid_to_has_ans, na_prob_thresh):
-    new_scores = {}
-    for qid, s in scores.items():
-        pred_na = na_probs[qid] > na_prob_thresh
-        if pred_na:
-            new_scores[qid] = float(not qid_to_has_ans[qid])
-        else:
-            new_scores[qid] = s
-    return new_scores
-
-
-def get_raw_scores(dataset, preds):
-    exact_scores = {}
-    f1_scores = {}
-    for article in dataset:
-        for p in article['paragraphs']:
-            for qa in p['qas']:
-                qid = qa['id']
-                gold_answers = [a['text'] for a in qa['answers'] if normalize_answer(a['text'])]
-                if not gold_answers:
-                    # For unanswerable questions, only correct answer is empty string
-                    gold_answers = ['']
-                if qid not in preds:
-                    print('Missing prediction for %s' % qid)
-                    continue
-                a_pred = preds[qid]
-                # Take max over all gold answers
-                exact_scores[qid] = max(compute_exact(a, a_pred) for a in gold_answers)
-                f1_scores[qid] = max(compute_f1(a, a_pred) for a in gold_answers)
-    return exact_scores, f1_scores
-
-
-def make_qid_to_has_ans(dataset):
-    qid_to_has_ans = {}
-    for article in dataset:
-        for p in article['paragraphs']:
-            for qa in p['qas']:
-                qid_to_has_ans[qa['id']] = bool(qa['answers'])
-
-    return qid_to_has_ans
-
-
-def evaluate_v2(dataset, predictions):
-    na_probs = {k: 0.0 for k in predictions}
-
-    qid_to_has_ans = make_qid_to_has_ans(dataset)  # maps qid to True/False
-    has_ans_qids = [k for k, v in qid_to_has_ans.items() if v]
-    no_ans_qids = [k for k, v in qid_to_has_ans.items() if not v]
-
-    exact_raw, f1_raw = get_raw_scores(dataset, predictions)
-
-    exact_thresh = apply_no_ans_threshold(exact_raw, na_probs, qid_to_has_ans, 1.0)
-    f1_thresh = apply_no_ans_threshold(f1_raw, na_probs, qid_to_has_ans, 1.0)
-
-    out_eval = make_eval_dict(exact_thresh, f1_thresh)
-
-    return out_eval
-
-
-def exact_match_score(prediction, ground_truth):
-    return (normalize_answer(prediction) == normalize_answer(ground_truth))
-
-
-def metric_max_over_ground_truths(metric_fn, prediction, ground_truths):
-    scores_for_ground_truths = []
-    for ground_truth in ground_truths:
-        score = metric_fn(prediction, ground_truth)
-        scores_for_ground_truths.append(score)
-    return max(scores_for_ground_truths)
-
-
-def evaluate(dataset, predictions):
-    f1 = exact_match = total = 0
-    for article in dataset:
-        for paragraph in article['paragraphs']:
-            for qa in paragraph['qas']:
-                total += 1
-                if qa['id'] not in predictions:
-                    message = 'Unanswered question ' + qa['id'] + \
-                              ' will receive score 0.'
-                    print(message, file=sys.stderr)
-                    continue
-                ground_truths = list(map(lambda x: x['text'], qa['answers']))
-                prediction = predictions[qa['id']]
-
-                exact_match += metric_max_over_ground_truths(
-                    exact_match_score, prediction, ground_truths)
-                f1 += metric_max_over_ground_truths(
-                    _f1_score, prediction, ground_truths)
-
-    exact_match = 100.0 * exact_match / total
-    f1 = 100.0 * f1 / total
-    return {'em': exact_match, 'f1': f1}
-
-
-def write_predictions(all_examples, all_features, all_results, n_best_size,
-                      max_answer_length, do_lower_case, output_prediction_file,
-                      output_nbest_file, output_null_log_odds_file, verbose_logging,
-                      version_2_with_negative, null_score_diff_threshold, dev_dataset, infer_times):
-    """Write final predictions to the json file and log-odds of null if needed."""
-    #     logger.info("Writing predictions to: %s" % (output_prediction_file))
-    #     logger.info("Writing nbest to: %s" % (output_nbest_file))
-
-    example_index_to_features = collections.defaultdict(list)
-    for feature in all_features:
-        example_index_to_features[feature.example_index].append(feature)
-
-    unique_id_to_result = {}
-    for result in all_results:
-        unique_id_to_result[result.unique_id] = result
-
-    _PrelimPrediction = collections.namedtuple(  # pylint: disable=invalid-name
-        "PrelimPrediction",
-        ["feature_index", "start_index", "end_index", "start_logit", "end_logit"])
-
-    all_predictions = collections.OrderedDict()
-    all_nbest_json = collections.OrderedDict()
-    scores_diff_json = collections.OrderedDict()
-
-    for (example_index, example) in enumerate(all_examples):
-        features = example_index_to_features[example_index]
-
-        prelim_predictions = []
-        # keep track of the minimum score of null start+end of position 0
-        score_null = 1000000  # large and positive
-        min_null_feature_index = 0  # the paragraph slice with min null score
-        null_start_logit = 0  # the start logit at the slice with min null score
-        null_end_logit = 0  # the end logit at the slice with min null score
-        for (feature_index, feature) in enumerate(features):
-            result = unique_id_to_result[feature.unique_id]
-            start_indexes = _get_best_indexes(result.start_logits, n_best_size)
-            end_indexes = _get_best_indexes(result.end_logits, n_best_size)
-            # if we could have irrelevant answers, get the min score of irrelevant
-            if version_2_with_negative:
-                feature_null_score = result.start_logits[0] + result.end_logits[0]
-                if feature_null_score < score_null:
-                    score_null = feature_null_score
-                    min_null_feature_index = feature_index
-                    null_start_logit = result.start_logits[0]
-                    null_end_logit = result.end_logits[0]
-            for start_index in start_indexes:
-                for end_index in end_indexes:
-                    # We could hypothetically create invalid predictions, e.g., predict
-                    # that the start of the span is in the question. We throw out all
-                    # invalid predictions.
-                    if start_index >= len(feature.tokens):
-                        continue
-                    if end_index >= len(feature.tokens):
-                        continue
-                    if start_index not in feature.token_to_orig_map:
-                        continue
-                    if end_index not in feature.token_to_orig_map:
-                        continue
-                    if not feature.token_is_max_context.get(start_index, False):
-                        continue
-                    if end_index < start_index:
-                        continue
-                    length = end_index - start_index + 1
-                    if length > max_answer_length:
-                        continue
-                    prelim_predictions.append(
-                        _PrelimPrediction(
-                            feature_index=feature_index,
-                            start_index=start_index,
-                            end_index=end_index,
-                            start_logit=result.start_logits[start_index],
-                            end_logit=result.end_logits[end_index]))
-        if version_2_with_negative:
-            prelim_predictions.append(
-                _PrelimPrediction(
-                    feature_index=min_null_feature_index,
-                    start_index=0,
-                    end_index=0,
-                    start_logit=null_start_logit,
-                    end_logit=null_end_logit))
-        prelim_predictions = sorted(
-            prelim_predictions,
-            key=lambda x: (x.start_logit + x.end_logit),
-            reverse=True)
-
-        _NbestPrediction = collections.namedtuple(  # pylint: disable=invalid-name
-            "NbestPrediction", ["text", "start_logit", "end_logit"])
-
-        seen_predictions = {}
-        nbest = []
-        for pred in prelim_predictions:
-            if len(nbest) >= n_best_size:
-                break
-            feature = features[pred.feature_index]
-            if pred.start_index > 0:  # this is a non-null prediction
-                tok_tokens = feature.tokens[pred.start_index:(pred.end_index + 1)]
-                orig_doc_start = feature.token_to_orig_map[pred.start_index]
-                orig_doc_end = feature.token_to_orig_map[pred.end_index]
-                orig_tokens = example.doc_tokens[orig_doc_start:(orig_doc_end + 1)]
-                tok_text = " ".join(tok_tokens)
-
-                # De-tokenize WordPieces that have been split off.
-                tok_text = tok_text.replace(" ##", "")
-                tok_text = tok_text.replace("##", "")
-
-                # Clean whitespace
-                tok_text = tok_text.strip()
-                tok_text = " ".join(tok_text.split())
-                orig_text = " ".join(orig_tokens)
-
-                final_text = get_final_text(tok_text, orig_text, do_lower_case, verbose_logging)
-                if final_text in seen_predictions:
-                    continue
-
-                seen_predictions[final_text] = True
-            else:
-                final_text = ""
-                seen_predictions[final_text] = True
-
-            nbest.append(
-                _NbestPrediction(
-                    text=final_text,
-                    start_logit=pred.start_logit,
-                    end_logit=pred.end_logit))
-        # if we didn't include the empty option in the n-best, include it
-        if version_2_with_negative:
-            if "" not in seen_predictions:
-                nbest.append(
-                    _NbestPrediction(
-                        text="",
-                        start_logit=null_start_logit,
-                        end_logit=null_end_logit))
-
-            # In very rare edge cases we could only have single null prediction.
-            # So we just create a nonce prediction in this case to avoid failure.
-            if len(nbest) == 1:
-                nbest.insert(0,
-                             _NbestPrediction(text="empty", start_logit=0.0, end_logit=0.0))
-
-        # In very rare edge cases we could have no valid predictions. So we
-        # just create a nonce prediction in this case to avoid failure.
-        if not nbest:
-            nbest.append(
-                _NbestPrediction(text="empty", start_logit=0.0, end_logit=0.0))
-
-        assert len(nbest) >= 1
-
-        total_scores = []
-        best_non_null_entry = None
-        for entry in nbest:
-            total_scores.append(entry.start_logit + entry.end_logit)
-            if not best_non_null_entry:
-                if entry.text:
-                    best_non_null_entry = entry
-
-        probs = _compute_softmax(total_scores)
-
-        nbest_json = []
-        for (i, entry) in enumerate(nbest):
-            output = collections.OrderedDict()
-            output["text"] = entry.text
-            output["probability"] = probs[i]
-            output["start_logit"] = entry.start_logit
-            output["end_logit"] = entry.end_logit
-            nbest_json.append(output)
-
-        assert len(nbest_json) >= 1
-
-        if not version_2_with_negative:
-            all_predictions[example.qas_id] = nbest_json[0]["text"]
-        else:
-            # predict "" iff the null score - the score of best non-null > threshold
-            score_diff = score_null - best_non_null_entry.start_logit - (
-                best_non_null_entry.end_logit)
-            scores_diff_json[example.qas_id] = score_diff
-            if score_diff > null_score_diff_threshold:
-                all_predictions[example.qas_id] = ""
-            else:
-                all_predictions[example.qas_id] = best_non_null_entry.text
-            all_nbest_json[example.qas_id] = nbest_json
-
-    with open(output_prediction_file, "w") as writer:
-        writer.write(json.dumps(all_predictions, indent=4) + "\n")
-    with open(output_nbest_file, "w") as writer:
-        writer.write(json.dumps(all_nbest_json, indent=4) + "\n")
-    if version_2_with_negative:
-        with open(output_null_log_odds_file, "w") as writer:
-            writer.write(json.dumps(scores_diff_json, indent=4) + "\n")
-
-    if version_2_with_negative:
-        result = evaluate_v2(dev_dataset, all_predictions)
-    else:
-        result = evaluate(dev_dataset, all_predictions)
-
-    result['infer_cnt'] = len(infer_times)
-    result['infer_time'] = sum(infer_times) / len(infer_times)
-    return result
-
-
-def do_qa_eval(args, model, dataloader, features, examples, device,
-               dev_dataset, subbert_config):
-    all_results = []
-    infer_times = []
-    for _, batch_ in enumerate(dataloader):
-        batch_ = tuple(t.to(device) for t in batch_)
-        input_ids, input_mask, segment_ids, example_indices = batch_
-        with torch.no_grad():
-            start = datetime.now()
-            (batch_start_logits, batch_end_logits) = model(input_ids, subbert_config,
-                                                           input_mask, segment_ids)
-            infer_times.append((datetime.now() - start).microseconds / 1000)
-        for i, example_index in enumerate(example_indices):
-            start_logits = batch_start_logits[i].detach().cpu().tolist()
-            end_logits = batch_end_logits[i].detach().cpu().tolist()
-            eval_feature = features[example_index.item()]
-            unique_id = int(eval_feature.unique_id)
-            all_results.append(RawResult(unique_id=unique_id, start_logits=start_logits, end_logits=end_logits))
-
-    output_prediction_file = os.path.join(OUTPUT_DIR, "predictions.json")
-    output_nbest_file = os.path.join(OUTPUT_DIR, "nbest_predictions.json")
-    output_null_log_odds_file = os.path.join(OUTPUT_DIR, "null_odds.json")
-
-    return write_predictions(examples, features, all_results, args.n_best_size, args.max_answer_length,
-                             args.do_lower_case, output_prediction_file,
-                             output_nbest_file, output_null_log_odds_file,
-                             args.verbose_logging, args.version_2_with_negative,
-                             args.null_score_diff_threshold, dev_dataset, infer_times)
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--data_dir",
-                        default=None,
-                        type=str,
-                        required=True,
-                        help="The input data dir. Should contain the .tsv files (or other data files) for the task.")
-    parser.add_argument("--model",
-                        default=None,
-                        type=str,
-                        required=True,
-                        help="The student model dir.")
-    parser.add_argument("--task_name",
-                        default=None,
-                        type=str,
-                        required=True,
-                        help="The name of the task to train.")
-    parser.add_argument("--cache_dir",
-                        default="",
-                        type=str,
-                        help="Where do you want to store the pre-trained models downloaded from s3")
-    parser.add_argument("--max_seq_length",
-                        default=128,
-                        type=int,
-                        help="The maximum total input sequence length after WordPiece tokenization. \n"
-                             "Sequences longer than this will be truncated, and sequences shorter \n"
-                             "than this will be padded.")
-    parser.add_argument("--do_lower_case",
-                        action='store_true',
-                        help="Set this flag if you are using an uncased model.")
-    parser.add_argument("--train_batch_size",
-                        default=32,
-                        type=int,
-                        help="Total batch size for training.")
-    parser.add_argument("--eval_batch_size",
-                        default=32,
-                        type=int,
-                        help="Total batch size for eval.")
-    parser.add_argument("--learning_rate",
-                        default=5e-5,
-                        type=float,
-                        help="The initial learning rate for Adam.")
-    parser.add_argument('--weight_decay', '--wd',
-                        default=1e-4,
-                        type=float,
-                        metavar='W',
-                        help='weight decay')
-    parser.add_argument("--num_train_epochs",
-                        default=3.0,
-                        type=float,
-                        help="Total number of training epochs to perform.")
-    parser.add_argument("--warmup_proportion",
-                        default=0.1,
-                        type=float,
-                        help="Proportion of training to perform linear learning rate warmup for. "
-                             "E.g., 0.1 = 10%% of training.")
-    parser.add_argument("--no_cuda",
-                        action='store_true',
-                        help="Whether not to use CUDA when available")
-    parser.add_argument('--seed',
-                        default="42",
-                        type=str,
-                        help="random seed for initialization")
-    parser.add_argument('--gradient_accumulation_steps',
-                        type=int,
-                        default=1,
-                        help="Number of updates steps to accumulate before performing a backward/update pass.")
-    parser.add_argument("--doc_stride", default=128, type=int,
-                        help="When splitting up a long document into chunks, how much stride to take between chunks.")
-
-    # SQuAD hyper-parameters
-    parser.add_argument("--n_best_size", default=20, type=int,
-                        help="The total number of n-best predictions to generate in the nbest_predictions.json "
-                             "output file.")
-    parser.add_argument("--max_answer_length", default=30, type=int,
-                        help="The maximum length of an answer that can be generated. This is needed because the start "
-                             "and end predictions are not conditioned on one another.")
-    parser.add_argument('--version_2_with_negative', default=0,type=int)     # 0 means 1.1, 1 means 2.0
-    parser.add_argument('--null_score_diff_threshold',
-                        type=float, default=0.0,
-                        help="If null_score - best_non_null is greater than the threshold predict null.")
-    parser.add_argument("--max_query_length",
-                        default=64,
-                        type=int,
-                        help="The maximum number of tokens for the question. Questions longer than this will "
-                             "be truncated to this length.")
-    parser.add_argument("--train_file", default='SQuAD2.0/train-v2.0.json', type=str,
-                        help="SQuAD json for training. E.g., train-v1.1.json")
-    parser.add_argument("--predict_file", default='SQuAD2.0/dev-v2.0.json', type=str,
-                        help="SQuAD json for predictions. E.g., dev-v1.1.json or test-v1.1.json")
-    parser.add_argument("--verbose_logging", action='store_true',
-                        help="If true, all of the warnings related to data processing will be printed. "
-                             "A number of warnings are expected for a normal SQuAD evaluation.")
-
-    # Evaluation of sub-model
-    parser.add_argument('--arches_file', default='', type=str)
-    parser.add_argument('--super_model', default='MLM', type=str)
-
-    # Run on modelarts
-    parser.add_argument("--data_url", type=str, default="", help="s3 url")
-    parser.add_argument("--train_url", type=str, default="", help="s3 url")
-    parser.add_argument("--init_method", default='', type=str)
-
-    args = parser.parse_args()
-
-    processors = {
-        "squad1": Squad1Processor
-    }
-
-    output_modes = {
-        "squad1": "qa_classification"
-    }
-
-    # fine-tuning default parameters
-    default_params = {
-        "squad1": {"num_train_epochs": 4, "max_seq_length": 384,
-                   "learning_rate": 3e-5, "eval_step": 50, "train_batch_size": 16},
-    }
-
-    #acc_tasks = ["mnli", "mrpc", "sst-2", "qqp", "qnli", "rte"]
-    #corr_tasks = ["sts-b"]
-    #mcc_tasks = ["cola"]
-    qa_tasks = ["squad1"]#, "squad2"]
-
-    # Prepare devices
-    # Always use the first gpu
-    os.environ["CUDA_VISIBLE_DEVICES"] = "0"        # TODO
-
-    device = torch.device("cuda" if torch.cuda.is_available() and not args.no_cuda else "cpu")
-    n_gpu = torch.cuda.device_count()
-
-    logging.basicConfig(format='%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
-                        datefmt='%m/%d/%Y %H:%M:%S',
-                        level=logging.INFO)
-
-    logger.info("device: {} n_gpu: {}".format(device, n_gpu))
-
-    def str2intlist(str):
-        strs = str.split()
-        intlist = list(map(int, strs))
-        return intlist
-
-    def set_seeds(seed, n_gpu):
-        random.seed(seed)
-        np.random.seed(seed)
-        torch.manual_seed(seed)
-        if n_gpu > 0:
-            torch.cuda.manual_seed_all(seed)
-
-    # set_seeds(args.seed, n_gpu)
-
-    seeds = str2intlist(args.seed)
-    task_lis = ['squad1']
-    data_dir_lis = ['/home/vmagent/app/dataset/SQuAD']
-    assert len(task_lis) == len(data_dir_lis)
-
-    subbert_configs = list()
-    subbert_config = dict()
-    tokenizer = BertTokenizer.from_pretrained(MODEL_DIR, do_lower_case=args.do_lower_case)
-
-    for seed in seeds:
-        set_seeds(seed, n_gpu)
-        config = BertConfig.from_pretrained(MODEL_DIR, num_labels=2, finetuning_task='sst-2')
-
-        if not args.arches_file:
-            subbert_config['sample_layer_num'] = config.num_hidden_layers
-            subbert_config['sample_hidden_size'] = config.hidden_size
-            subbert_config['sample_intermediate_sizes'] = [config.intermediate_size] * config.num_hidden_layers
-            subbert_config['sample_num_attention_heads'] = [config.num_attention_heads] * config.num_hidden_layers
-            try:
-                subbert_config['sample_qkv_sizes'] = [config.qkv_size] * config.num_hidden_layers
-            except:
-                subbert_config['sample_qkv_sizes'] = [config.hidden_size] * config.num_hidden_layers
-            logger.info('subbert_config: {}'.format(subbert_config))
-            subbert_configs.append(subbert_config)
-        else:
-            subbert_configs = []
-            base_arches_file = os.path.abspath(os.path.dirname(os.path.dirname(__file__)))
-            with open(os.path.join(base_arches_file, "best_model_structure.txt"), 'r') as fin:
-                for line in fin:
-                    line = line.strip()
-                    subbert_configs.append(get_subconfig(ast.literal_eval(line)))
-
-        print('subbert_configs: {}'.format(subbert_configs))
-        #OUTPUT_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)), OUTPUT_DIR)
-
-        for subbert_config in subbert_configs:
-            for task_name, data_dir in zip(task_lis, data_dir_lis):
-                if task_name == 'squad1':
-                    train_file = os.path.join(DATA_DIR, 'train-v1.1.json')
-                    predict_file = os.path.join(DATA_DIR, 'dev-v1.1.json')
-
-                # Prepare task settings
-                # if not oncloud and os.path.exists(OUTPUT_DIR) and os.listdir(OUTPUT_DIR):
-                #     raise ValueError("Output directory ({}) already exists and is not empty.".format(OUTPUT_DIR))
-                if not os.path.exists(OUTPUT_DIR):
-                    os.makedirs(OUTPUT_DIR)
-
-                task_name = task_name.lower()
-                if task_name in default_params:
-                    args.num_train_epochs = default_params[task_name]["num_train_epochs"]
-                    args.max_seq_length = default_params[task_name]["max_seq_length"]
-                    args.learning_rate = default_params[task_name]["learning_rate"]
-                    args.eval_step = default_params[task_name]["eval_step"]
-
-                    if 'train_batch_size' in default_params[task_name]:
-                        args.train_batch_size = default_params[task_name]["train_batch_size"]
-
-
-                if task_name not in processors:
-                    raise ValueError("Task not found: %s" % task_name)
-
-                processor = processors[task_name]()
-                output_mode = output_modes[task_name]
-                label_list = processor.get_labels()
-                num_labels = len(label_list)
-
-                config.num_labels = num_labels
-                config.finetuning_task = task_name
-
-                train_data, train_examples = read_squad_examples(
-                    input_file=train_file, is_training=True,
-                    version_2_with_negative=args.version_2_with_negative)
-
-                if args.gradient_accumulation_steps < 1:
-                    raise ValueError("Invalid gradient_accumulation_steps parameter: {}, should be >= 1".format(
-                        args.gradient_accumulation_steps))
-
-                args.train_batch_size = args.train_batch_size // args.gradient_accumulation_steps
-
-                num_train_optimization_steps = int(
-                    len(train_examples) / args.train_batch_size / args.gradient_accumulation_steps * args.num_train_epochs)
-
-                def get_dataloader(examples, label_list, max_seq_length, tokenizer, output_mode,
-                                   is_dev=False, is_training=False):
-                    features = qa_convert_examples_to_features(
-                        examples=examples,
-                        tokenizer=tokenizer,
-                        max_seq_length=args.max_seq_length,
-                        doc_stride=args.doc_stride,
-                        max_query_length=args.max_query_length,
-                        is_training=is_training)
-
-                    all_input_ids = torch.tensor([f.input_ids for f in features], dtype=torch.long)
-                    all_input_mask = torch.tensor([f.input_mask for f in features], dtype=torch.long)
-                    all_segment_ids = torch.tensor([f.segment_ids for f in features], dtype=torch.long)
-                    if is_training:
-                        all_start_positions = torch.tensor([f.start_position for f in features], dtype=torch.long)
-                        all_end_positions = torch.tensor([f.end_position for f in features], dtype=torch.long)
-                        data = TensorDataset(all_input_ids, all_input_mask, all_segment_ids,
-                                             all_start_positions, all_end_positions)
-                    else:
-                        all_example_index = torch.arange(all_input_ids.size(0), dtype=torch.long)
-                        data = TensorDataset(all_input_ids, all_input_mask, all_segment_ids, all_example_index)
-
-                    if not is_dev:
-                        sampler = RandomSampler(data)
-                    else:
-                        sampler = SequentialSampler(data)
-                    return features, DataLoader(data, sampler=sampler, batch_size=args.train_batch_size)
-
-                labels, train_dataloader = get_dataloader(train_examples, label_list, args.max_seq_length,
-                                                          tokenizer, output_mode, is_training=True)
-                
-                eval_dataset, eval_examples = read_squad_examples(input_file=predict_file, is_training=False,
-                                                                    version_2_with_negative=args.version_2_with_negative)
-
-                eval_features, eval_dataloader = get_dataloader(eval_examples, label_list, args.max_seq_length,
-                                                                tokenizer, output_mode, is_dev=True, is_training=False)
-
-                # Prepare loss functions
-                loss_mse = MSELoss()
-                loss_fct = CrossEntropyLoss()
-
-                model = SuperBertForQuestionAnswering.from_pretrained(MODEL_DIR, config)
-                model.to(device)
-
-                if n_gpu > 1:
-                    model = torch.nn.DataParallel(model)
-
-                model.module.set_sample_config(subbert_config) if hasattr(model, 'module') \
-                    else model.set_sample_config(subbert_config)
-                # Prepare optimizer
-                param_optimizer = list(model.named_parameters())
-                size = 0
-                for n, p in model.named_parameters():
-                    size += p.nelement()
-                logger.info('Total parameters: {}'.format(size))
-
-                no_decay = ['bias', 'LayerNorm.bias', 'LayerNorm.weight']
-                optimizer_grouped_parameters = [
-                    {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay)],
-                     'weight_decay': 0.01},
-                    {'params': [p for n, p in param_optimizer if any(nd in n for nd in no_decay)], 'weight_decay': 0.0}
-                ]
-                schedule = 'warmup_linear'
-                optimizer = BertAdam(optimizer_grouped_parameters,
-                                     schedule=schedule,
-                                     lr=args.learning_rate,
-                                     warmup=args.warmup_proportion,
-                                     t_total=num_train_optimization_steps)
-
-                logger.info("***** Running training *****")
-                print("  Num examples = %d" % (len(train_examples)))
-                print("  Batch size = %d" % (args.train_batch_size))
-                print("  Num steps = %d" % (num_train_optimization_steps))
-
-                # Train and evaluate
-                global_step = 0
-                best_dev_acc = 0.0
-                best_dev_acc_str = ''
-                infer_cnt = 0
-                infer_times = []
-                output_eval_file = os.path.join(OUTPUT_DIR, "eval_results.txt")
-
-                start_time = time.time()
-                for epoch_ in trange(int(args.num_train_epochs), desc="Epoch"):
-                    tr_loss = 0.
-                    tr_cls_loss = 0.
-
-                    model.train()
-                    nb_tr_examples, nb_tr_steps = 0, 0
-
-                    for step, batch in enumerate(tqdm(train_dataloader, desc="Iteration", ascii=True)):
-                        batch = tuple(t.to(device) for t in batch)
-
-                        input_ids, input_mask, segment_ids, start_positions, end_positions = batch
-
-                        if input_ids.size()[0] != args.train_batch_size:
-                            continue
-
-                        cls_loss = 0.
-
-                        cls_loss = model(input_ids, subbert_config, input_mask,
-                                            segment_ids, start_positions, end_positions,
-                                            kd_infer=True if args.super_model == 'KD' else False)
-
-                        loss = cls_loss
-                        tr_cls_loss += cls_loss.item()
-
-                        if n_gpu > 1:
-                            loss = loss.mean()  # mean() to average on multi-gpu.
-                        if args.gradient_accumulation_steps > 1:
-                            loss = loss / args.gradient_accumulation_steps
-
-                        loss.backward()
-
-                        tr_loss += loss.item()
-                        nb_tr_examples += input_ids.size(0)
-                        nb_tr_steps += 1
-
-                        if (step + 1) % args.gradient_accumulation_steps == 0:
-                            optimizer.step()
-                            optimizer.zero_grad()
-                            global_step += 1
-
-                        if (global_step + 1) % args.eval_step == 0:
-                            logger.info("***** Running evaluation *****")
-                            logger.info("  Epoch = {} iter {} step".format(epoch_, global_step))
-                            logger.info("  Num examples = %d", len(eval_examples))
-
-                            model.eval()
-
-                            loss = tr_loss / (step + 1)
-                            cls_loss = tr_cls_loss / (step + 1)
-
-                            result = do_qa_eval(args, model, eval_dataloader, eval_features, eval_examples,
-                                                    device, eval_dataset, subbert_config)
-
-                            infer_cnt += result['infer_cnt']
-                            infer_times.append(result['infer_time'])
-
-                            result['global_step'] = global_step
-                            result['cls_loss'] = cls_loss
-                            result['loss'] = loss
-
-                            result_to_file(result, output_eval_file)
-
-                            update_best = False
-
-                            if task_name in qa_tasks and result['f1'] + result['em'] > best_dev_acc:
-                                best_dev_acc = result['f1'] + result['em']
-                                best_dev_acc_str = 'f1: {}; em: {}'.format(result['f1'], result['em'])
-                                update_best = True
-
-                            if update_best:
-                                # Save a trained model
-                                model_name = "{}_seed{}".format(task_name, str(seed))
-                                logging.info("** ** * Saving fine-tuned model ** ** * ")
-                                # Only save the model it-self
-                                model_to_save = model.module if hasattr(model, 'module') else model
-                                output_model_file = os.path.join(OUTPUT_DIR, model_name)
-                                torch.save(model_to_save.state_dict(), output_model_file)
-                                tokenizer.save_vocabulary(OUTPUT_DIR)
-
-                            model.train()
-
-                end_time = time.time()
-                model_to_save = model.module if hasattr(model,
-                                                        'module') else model
-                parameter_size = model_to_save.calc_sampled_param_num()
-
-                output_str = "**************S*************\n" + \
-                             "task_name = {}\n".format(task_name) + \
-                             "architecture = {}\n".format(subbert_config) + \
-                             "parameter size = {}\n".format(parameter_size) + \
-                             "total training time = {}\n".format(str(end_time-start_time)) + \
-                             "best_acc = %s\n" % best_dev_acc_str + \
-                             "time_per_batch_infer = %.3f ms\n" % (sum(infer_times) / len(infer_times)) +\
-                             "infer_cnt = %d\n" % infer_cnt +\
-                             "**************E*************\n"
-
-                print(output_str)
-                output_eval_file = os.path.join(OUTPUT_DIR, "subbert.results")
-                with open(output_eval_file, "a+") as writer:
-                    writer.write(output_str + '\n')
-
-
-if __name__ == "__main__":
-    main()
